+++ bin/.gitignore
@@ -1,0 +1,3 @@
+/Renderer.class
+/Display.class
+/Graphics.class
